pyproject.toml
[project]
name = "edulance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
]
